walletrpc: add raw_tx field to BumpFee response
This commit adds an optional include_raw_tx field to BumpFeeRequest and a raw_tx field to BumpFeeResponse. When include_raw_tx is set to true, the RPC will wait for the sweep transaction to be created and return its hex-encoded raw transaction.

This allows users to manually broadcast transactions in cases where they don't propagate automatically, addressing issue #8470.

Changes:
- Add include_raw_tx bool field to BumpFeeRequest proto
- Add raw_tx string field to BumpFeeResponse proto
- Modify BumpFee implementation to optionally wait for sweep result
- Update sweepNewInput to return result channel
- Add --include_raw_tx flag to lncli wallet bumpfee command

Closes #8470